refactor(#3037): batch non-hot service→server backflow relocations (33→27)
Two coherent, behavior-preserving clusters that lower service→server
backflow without touching the reserved hot files.

Cluster A — stale-facade re-routes (services symbols imported via server path):
- `adk_session.rs` (2 sites): `HookSessionBody` real home is
  `crate::services::dispatched_sessions`; rewrite the two construction sites
  to import it directly instead of through
  `crate::server::routes::dispatched_sessions::*`.
- `runtime_bootstrap.rs` (1 site): `gc_stale_fixed_working_sessions_db_pg`
  real home is `crate::db::dispatched_sessions`; call it directly. Removed the
  now-orphaned `pub use` facade in `server/routes/dispatched_sessions.rs`.

Cluster B — kanban auth/identity helpers relocated to the services layer:
- Moved `require_explicit_bearer_token`, `resolve_requesting_agent_id_with_pg`
  (and their private helpers `trimmed_header_value`,
  `resolve_agent_id_from_channel_id_with_pg`) from `server/routes/kanban.rs`
  to `services/kanban_cards.rs`. They depend only on config + db (lower
  layers). `server/routes/kanban.rs` now imports them via the services facade
  (`crate::services::kanban::*`, forward dependency). Updated the 3
  `services/auto_queue/{control,order}_routes.rs` call sites to the services
  path (3 backflow findings removed).

Behavior unchanged (pure relocation, no logic edits). No circular deps. All
call sites updated. Facade re-exports kept only for still-consumed symbols
(no unused-import warnings).

Reserved hot files left untouched for the separate hot-file PR (6 findings):
turn_bridge/completion_guard.rs (5), tmux_watcher.rs (1). Also left:
dto/agents.rs sites (agents/query.rs, agents/turn.rs — #3238 in flight),
plus the larger thread_reuse / ws-event-bus / AppState / escalation-settings
relocations (broad churn, deferred to dedicated PRs).

Gates: cargo fmt --check clean; cargo check --lib and --lib --tests clean
(zero new warnings); touched-module tests pass; backflow baseline lowered
33→27; inventory + agent-maintenance freshness pass; ci-script-checks exit 0.
